+++ tests/addition.cpp
@@ -58,12 +58,12 @@ TEST_CASE("Negative Addition") {
 
 TEST_CASE("Random Generation") {
     std::mt19937_64 rng;
-    std::uniform_int_distribution<int64_t> distribution;
+    std::uniform_int_distribution<long long int> distribution;
 
     for (int i{0}; i < NUM_GENERATIONAL_TESTS; ++i) {
-        const int64_t n1{distribution(rng) / 2};
-        const int64_t n2{distribution(rng) / 2};
-        const int64_t sum{n1 + n2};
+        const long long int n1{distribution(rng) / 2};
+        const long long int n2{distribution(rng) / 2};
+        const long long int sum{n1 + n2};
 
         BigInt big1{n1};
         BigInt big2{n2};
